Revert "Refactor restructure command handlers"
