+++ README.md
@@ -45,19 +45,6 @@
  
 ## 📦 更新记录
 
-#### v1.7.3 (命令注册与架构重构，不改变游戏逻辑)
-
-- 主入口重构：
-  - 统一采用方法装饰器（@filter.command）注册命令，移除类属性式绑定，符合官方规范，确保命令被框架正确扫描与注册。
-  - `main.py` 仅作为分发入口，命令实现全部下沉到 `handlers/*`，结构更清晰、可维护性更高。
-- 导入与目录：
-  - 使用精确导入的仓储与服务实现（`core.repositories.*`、`core.services.*`），移除通配导入，避免潜在的运行时隐患。
-- 命令与别名：
-  - 新增别名：「我的上架」支持「我的挂单」。
-  - 「驱灵」命令的实现函数名调整为 `dispel_protection`（命令名称与用法不变）。
-- 向后兼容：
-  - 不改变任何业务逻辑与数据结构；现有存档与指令行为保持不变。
-
 #### v1.7.2 (反制海灵守护道具系统)
 
 - **新增反制道具系统**：
@@ -360,7 +347,6 @@
 |---|---|
 | `/排行榜` | 查看金币排行榜 |
 | `/偷鱼 [@用户]` | 偷取指定用户的一条鱼 |
-| `/驱灵 [@用户]` | 驱散目标的海灵守护（需持有驱灵香） |
 | `/查看称号` | 查看我拥有的称号 |
 | `/使用称号 [ID]` | 装备指定ID称号 |
 | `/查看成就` | 查看我的成就进度 |
