+++ draw/help.py
@@ -169,7 +169,7 @@ def draw_help_image():
     market = [
         ("全部卖出", "一键卖出\n鱼塘所有鱼"),
         ("保留卖出", "卖出所有鱼\n但每种保留一条"),
-        ("砸锅卖铁", "危险操作！清空\n全部鱼(非用/保)鱼竿饰品"),
+        ("砸锅卖铁", "危险操作！\n清空全部鱼(非在用/非保护)鱼竿和饰品"),
         ("出售稀有度 [1-5]", "卖出指定\n稀有度的鱼"),
         ("出售鱼竿 [ID]", "出售指定ID鱼竿"),
         ("出售所有鱼竿", "一键出售所有\n(非在用/非保护)鱼竿"),
@@ -199,7 +199,6 @@ def draw_help_image():
     social = [
         ("排行榜", "查看金币排行榜"),
         ("偷鱼 [@用户]", "偷取指定用户\n的一条鱼"),
-        ("驱灵 [@用户]", "驱散目标的\n海灵守护"),
         ("查看称号", "查看我拥有的称号"),
         ("使用称号 [ID]", "装备指定ID称号"),
         ("查看成就", "查看我的成就进度"),
