+++ metadata.yaml
@@ -1,5 +1,5 @@
 ﻿name: Fishing Another!
-version: 1.7.3
+version: 1.7.2
 author: "xxlemon"
 description: "升级版的钓鱼插件，附带后台管理界面（个性化钓鱼游戏！支持原版fashing2 1.4.6版本fishing存档继续游玩）"
 repo: https://github.com/xxlemon-io/astrbot_plugin_fishing
